pyproject.toml
[tool.poetry]
name = "drfbooking"
version = "0.1.0"
description = "Django DRF project"
authors = ["MaximBrezhnev <max.b04.03@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = ">=5.2.4,<6.0.0"
djangorestframework = "^3.16.0"
dynaconf = {extras = ["yaml"], version = "^3.2.11"}
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"
ruff = "^0.12.8"


[tool.poetry.group.test.dependencies]
faker = "^37.5.3"
pytest = "^8.4.1"
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
lint.extend-select = ["I"]
fix = true

[tool.ruff.lint.isort]
known-first-party = ["booking", "room"]
combine-as-imports = true
force-single-line = false